MVP Architecure in GenerateCollectionSheetFragment
