+++ mifosng-android/src/main/java/com/mifos/api/services/CenterService.java
@@ -40,9 +40,8 @@ public interface CenterService {
             (@Path("centerId") int centerId);
 
     @GET(APIEndPoint.CENTERS)
-    void getAllCentersInOffice(@Query("officeId") int officeId, @QueryMap Map<String, Object>
-            additionalParams,
-                               Callback<List<Center>> centersCallback);
+    Observable<List<Center>> getAllCentersInOffice(@Query("officeId") int officeId,
+                                                   @QueryMap Map<String, Object> additionalParams);
 
     @GET(APIEndPoint.CENTERS + "/{centerId}?associations=groupMembers")
     Observable<CenterWithAssociations> getAllGroupsForCenter(@Path("centerId") int centerId);
